refactor toll chat to be able to switch between osc and chat bot components
