+++ src/components/TollChat/ChatInterface.module.scss
@@ -437,10 +437,3 @@
     line-height: 1.3;
     color: #a61010
 }
-
-.callbackIframe {
-    display: none;
-    height: 0;
-    width: 0;
-    opacity: 0;
-}
